### Create a detail view layout Part 2
- Finishing DetailView by creating InfoBox. User can unfold the desired InfoBox.
DisclosureGroup view presents a disclosure indicator and content inside it.

User can find an indicator called Chevron on the opposite side of the title. Content is hidden inside

`ForEach(0..<nutrients.count, id: \.self) { item in {`

will repeat the content as many times as many elements are in the array of the nutrients

Adding `Divider().padding(.vertical, 2)` to divide Array vertically.

and Finally and `FruitsNutrientsView(fruit: fruit)` to the FruitDetailView

- Always need to add the necessary parameter for it since we are fetching information from the local datafile