Create Reset features - Create new property @AppStorage("isOnboarding") var isOnboarding: Bool = false
This code will create a new toggle with label.   Toggle(isOn: $isOnboarding) {

Added new user interface element to the layout

Use $isOnBoarding binding property

Tied together a switch on and off cases to that is on boarding property.
added a false value and the initialize property, so the initial state of toggle is off
when User tab on toggle, then its value will switch to true and the app storage
When the storage detects the value of change in this property, then it will invalidate the screen immediately.